Refactor GeneralConfig (now "Settings")
